Implement "Scale handles proportionally" feature for the mesh transform
When the user moves a node, the handles of the neighbouring nodes may
overlap. To avoid that we should shrink the size of the handles when
the distance between two nodes decreases.

The patch implements an option for the mesh transform tool that
activates such scaling (disabled by default).

This feature cannot go into 4.4.2 because of the string freeze.

CC:kimageshop@kde.org

# Conflicts:
#	plugins/tools/tool_transform2/tool_transform_args.h