Let's not crash on incorrect files in resource folder
Before this commit, Folder Iterator would crash on files
that don't have a proper loader. Since we never know what files
users can have, and the FolderStorage doesn't check for correctness,
let's not assert here.
This commit changes Q_ASSERT to warnKrita and returning false.

This happened to me on layerstyles/..._style file, which
doesn't have any mimetype nor any loader, because it was empty file,
because it was a failed attempt at saving an ephemeral resource.
However from my understanding of the code, it can happen for other
invalid resources, so still, let's not assert.