Don't add inactive versions of resources as new resources
Before this commit, if you changes the palette, it would all work
fine (all changes would result in a new version of the palette,
not a whole new resource), but after restart all the other versions
of the palette would reappear in the palette docker. That was caused
by the fact that ResourceCacheDb never checked whether the resource
is already a version of another resource when synchronizing database.

This commit fixes it by ensuring that 'resourceIdForResource' finds
the resource id of the current version of the resource even for the
resources that are marked as old versions of the resource (so they are
not present in the `resources` table but only in `versioned_resources`.

Possible problem: KisResourceCacheDb, noticing the resource already
exists, checks for timestamp and adds a new version if needed.
However at this point it *shouldn't* be needed, ever, no matter what
timestamps say.
If a filepath is in the database, then it means Krita knows about it
and already has this filepath remembered. Adding a new version with
the same filepath doesn't sound reasonable. If the user replaced one
resource with the same, called the same way, it means we might need
to update the thumbnail and/or the name, but adding anything to the
database - no, because there is no *more* resources, they were
*replaced*.
With proper versioning and the feature "revert to previous version",
it will lead to even more trouble, with the newest-by-timestamp
version replacing the user-selected current version on every restart.