Rename the preset instead of creating a copy
Before this commit, Krita would try to make a copy of the preset
with a new name and remove the older one. This commit uses the
KisResourceModel::renameResource() function instead,
which results in a renaming happening only in the database.

Note: because the resource is not valid after apparent changes,
renaming the resource still doesn't work (it returns from the
function early). However that should ensure no crashes.