Don't add versions of resource already present in the database
Problem: (copied from the commit with the title:
"Don't add inactive versions of resources as new resources"
(I cannot say the git hash because it's only local here
for now, so it would be incorrect anyway)

---
KisResourceCacheDb, noticing the resource already
exists, checks for timestamp and adds a new version if needed.
However at this point it *shouldn't* be needed, ever, no matter what
timestamps say.

If a filepath is in the database, then it means Krita knows about it
and already has this filepath remembered. Adding a new version with
the same filepath doesn't sound reasonable. If the user replaced one
resource with the same, called the same way, it means we might need
to update the thumbnail and/or the name, but adding anything to the
database - no, because there is no *more* resources, they were
*replaced*.

With proper versioning and the feature "revert to previous version",
it will lead to even more trouble, with the newest-by-timestamp
version replacing the user-selected current version on every restart.
---

This resulted in multiple versions with -1 number and generally
duplication of versions.

Instead, we should just update the timestamp. (Also why would we use
timestamp now, anyway?) Note: this commit doesn't do that yet.