Comment out the "Could not find..." debug again
Before this commit, the qWarning() "Could not find resource in the
database" warning was not commented out. That results in lots of
debug code that sounds scary but it actually happens on every resource
that is found in the local resources folder but not in the database
- which means for example on creating the database from scratch.
This commit comments it out again and adds a little comment
explaining why it's commented out.