Prevent infinite recursion with Qt >5.15.2
The fix for QTBUG-87781 is to recreate an internal mapping in
QSortFilterProxyModel::setSourceModel(), and KisResourceModel calls
this in the constructor.
This leads to KisResourceModel::filterAcceptsRow() being called, which
ultimately fetches data via KisResourceQueryMapper::variantFromResourceQuery().
So the latter must not create a temporary KisResourceModel when processing the
data role used by filterAcceptsRow().

So only create the temporary model when actually required.