Properly fix GMic layer naming
Turns out, GMic layer parsing is a three-stage process:

1. Image size is synced.
2. Layer count is synced; new layers are added if necessary (!!!)
3. Untouched layers are visited and updated with the data from GMic

While (3) is not much of an issue (after I added the name parsing code),
with (2) it means that some nodes will be overwritten, and its names
lost.

This commit fixes the issue by applying 1:1 the original name to the new
layer.