Fix generation of md5sum for the resources
We shouldn't re-save the resource to get calculate MD5, it will
generate random md5sum depending on the version of the underlying
libraries (say, libpng). Instead, we should calculate MD5 directly
in the storage.

This patch also caches MD5 in the storage, so that it shouldn't be
recalculated every time resource is loaded into Krita.