Add a Save button in Palette docker to limit number of versions
Before this commit, Palette docker would save a new version of the
palette every time anything changed on it. This commit
saved the palette to file only if the user clicks on "Save"
button.

Note: the palette is also saved with changes into the document without
the need to click "Save", and there is no way to reset the palette to
the previous version (neither for documents-embedded palettes nor
for folder-stored palettes).
This might look like inconsistency.

Another note: there is no visual way to indicate that Save needs to be
done because the palette has changed.