macOS: Finder plugins must use the loader_path of Krita.app
Xcode is picking up Quazip's requirement of our own bundled zlib.
Because the resulting plugins are not marked as executable, they are not
fixed by osxdeploy.sh. Moreover, the executables lack a rpath at all.

This commit marks the plugins' executables as +x and specifies a rpath
pointing to Krita's Frameworks folder. I also left an alternative option
inside osxdeploy.sh, if it is required in the future.

BUG: 432328