Don't save duplicate color spaces into Kpl
Before this commit, it was possible for the KoColorSet
to try to save multiple color profiles with the same name.
Since the Kpl format doesn't allow it anyway (the filename
is the id of the profile), we can assume that the profile
with the same name are just the same profile and not add
them to the list.
This commit ensures that only profiles with different names
are added to the list and saved into the device.