Workaround recursive calls in UndoStack
In some cases, when the user clicks in the history docker too
quickly, the undo requests may arrive recursively. Because there
are a few commands that demand processEvent() calls after their
execution. This patch workarounds the recursive issue using the
ticket-like system for the requests and ensuring that all of them
are executed at the topmost level of recursion.

CCBUG:432329