Android: provide host so pathPattern could be effective
Because host wasn't provided pathPattern would be ignored so Krita
showed up in app drawer with ability to open files of any possible
extension. With this the pathPattern *is* taken into consideration.
However, the only apps which I found consistent with pathPattern were
the default Files app on AOSP and Files by Google app.