Don't show changes in palettes embedded in documents
Before this commit, palettes embedded in the document would
still show changes even though they are only saved on
saving the document.
This commit just ignore the changes and always paint the
name of the palette in a way that suggests no changes have been made.
This is an easier and less complex solution than trying to
keep track of whether the document has been saved or not.