Fix opening brushes created during incomplete lightness feature implementation
We had a period of time, when the presence or absence of **color**
in the brush (not transparency) was used to decide whether to use
useColorAsAlpha. Basically, absence of color in the brush would switch
mask mode even when useColorAsAlpha was false.

In the new versions of Krita 'useColorAsAlpha' tag is completely
removed, and 'brushApplication' tag is used instead, which does
allow such ambiguity.

BUG:431937

# Conflicts:
#	libs/brush/kis_brushes_pipe.h
#	libs/brush/kis_imagepipe_brush.cpp