Add unittests for XML, change LAB handling for XML, handle old palettes.
1. Our XML serialization did have unittests, but only for roundtrips.
This patch adds unittests based on existing serializations, so we get
warned whenever we cannot load these properly.

2. Lab was set to 0-1, but should actually be L = 0 - 100 (inifinite),
a & b = -128 to 127 (infinite to both sides). This has been fixed now,
and there's a ton of unittests to make sure we're loading the right thing.

As every other LAB definition implementation uses this scheme, on top of
the original create swatch spec requiring these ranges, I think we'll avoid
a lot of headaches in the future by switching to this.

3. Our colorsets are now at version 2, and will convert old lab swatches
appropriately.

4. Also documents floating point LAB a bit better.

5. CMYK F32 is still weird, but I am at loss what is causing it.