Revise the fix for "Prevent infinite recursion with Qt >5.15.2"
Commit d062760c7be87da26f07f2111ba0b2cc55b29e4a prevents entering an
infinite recursion, but if QSortFilterProxyModel does any query on
KisAllResourceModel (as part of setting it as source model in
KisResourceModel), this changes the seek state of query object held
by the KisAllResourceModel singleton.

This query object however may be the input of
KisResourceQueryMapper::variantFromResourceQuery(), so the 'query'
parameter may silently change when creating a KisResourceModel there.

The desired information can be accessed more directly, though.