Update TestAsl/AbrStorage to reflect previous changes
Before this commit, Krita would fail some tests for storages
because of the recent changes to what the database would
return for various specified resource types.
This commit ensures that the test both checks if the list of resources
is returned when the proper resource type is submitted, and if the list
of the resources if empty when the improper type is submitted.