List resources by storage id, not storage type
Before this commit, Krita would list all resources from all
storages as long as the storage type matched. This was incorrect
because it was possible for multiple storages have the same type.
(Example: bundle, asl, abr)
This commit ensures that the storage has proper storage id and
that the "resourcesInDatabase" list only lists resources
that do belong in the specified storage.