Don't put resources updated on disk as new versions (again...)
Before this commit, Krita would try to add a resource that was updated
on the disk (= had newer timestamp) as a new version of the resource.
Since every database entry in versioned_resource should have an
equivalent in the storage, it makes sense only on the surface.
We shouldn't add any new entries: instead, in such case Krita should
update the entry in the database to reflect the changes
(update the thumbnail, timestamp, possibly name etc.).

Note: updateResource() function creates a new version, which means it
creates a new databas entry, which is different from updating the entry.
Hence we shouldn't use that one.

This commit ensures that synchronizeDb() code knows that two resource
entries (from database and storage) are the same if both resourceId
and version are the same, ignoring timestamps. That ensures that nothing
tries to add new versioned_resources entries.

TODO: the update of the database entry is not implemented yet.
It needs to recognize this situation in the if statement
in KisResourceCacheDb::synchronizeDb after the comment that starts with
"/// Now both arrays are sorted in resourceId/version/timestamp"
(line 1519),
and then make sure that the thumbnails etc. will be updated.
(Note: it should be quite similar to
KisResourceCacheDb::makeResourceTheCurrentVersion but possibly
with both versioned_resource and resources table?)