Handle starting Krita without any bundles or presets available
If there are no presets in the folder, and no bundles enabled,
krita will complain and open the bundle manager.

If you don't select a bundle that has presets in the bundle
manager, you won't be able to close the bundle manager dialog
until you've imported or enabled a bundle that has presets.

If you disable the last bundle in the bundle chooser in the
resource chooer, and there are no presets available, you will
get a critical warnign dialog

You can no longer disable the folder storage or the memory
storages in the bundle chooser

Note: Krita will still crash if there are no presets and you try
to paint, but you can start krita and even create a document.