Generalized popup widget.
We've taken the concept of the popup palette and generalized it
by allowing all of Krita's individual tools to return a popup
widget to be temporarily displayed directly on the canvas.

Ideally, this would work for menus, tool options, editor windows
and (of course) custom widgets like the popup palette. :)