Fix database to hav only unique tag to resource connection
Before this commit, the database could have multiple entries
in resource_tags table with the same resource_id and tag_id.
Since all the entries are just that, it makes no sense to allow that.
This commit adds a UNIQUE constraint on those two columns together
in the database in resource_tags table.