Use resourceId instead of resource in OneResController
Before this commit, TagSelectionWidgetOneResourceController
would use a whole KisResourceSP instead of resourceId.
This commit ensures the controller uses resourceId instead.

Before this commit, the model would get a whole resource which meant
that the resource had to be loaded first. This commit removes that need
by using only the resourceId for those actions.