Fix resource thumbnail preview size to be 100x100
Most brush presets thumbnails have size 200x200, and having size
128x128 would make the preview on 4k screen be in size 256x256.
That caused a need to scale the thumbnail.
Changing the size to 100x100 doesn't change much for the user - it's
still a big preview - but it makes it look better in the default
situation (for example, no need to upscale on 4k screen).