Fix not finding tags in Resource Manager in some situations
Before this commit, if the user clicked on some brushes (brush tips),
and then went to Brush Presets and clicked on some presets,
it would look as if there are no tags associated with those presets.
This commit fixes it by ensuring there will be a proper tag resources
model for a specific resource type.