Fix shortcuts with special keys on non-US layouts on Linux
WARNING: this patch solidifies the precedence of the us-latin layout
key positions over the current user layout. It basically means that
if the user's layout swaps the position of Z and Y letters, then Ctrl+Y
will still work as Ctrl+Z.

I need user's input on that change. Afaict, this bahavior was the default
on Windows, but it was not so on Linux before.

CCBUG:430479