Add create new tag widget inside the AddTag button
Before this commit, there was no way to create a new tag in the Resource
Manager. This commit adds a simple textbox with a [+] button at the end
at the end of the context menu that appears when you click on the AddTag
button.

There are still several issues regarding this addition.
- the popup only closes after the second Enter... it should close
always after the user presses Enter or clicks on the [+] button
(which doesn't happen)
- it adds QAction::trigger() in a class that is used elsewhere;
must check if it still works properly in the old location
(right-click in Brush Presets docker)
- KisSignalCompressor is most probably unnecessary