Fix crash in KisAllFilterTest caused by resource rewrite
Before this commit, KisAllFilterTest would first crash on
loading default configuration (which was fixed by using
KISTEST instead of QTEST) and then it would assert
because the global resources interface wasn't converted
to local resources interface which is what filters expect.

This commit doesn't mark the unit tests as working because
they still fail in multiple filters because of differences
between expected and actual results.

(Actual results often look similar, as if wrong configuration was
used).