Implemented linearization for icc incompatible transfer functions in heif.
ICC cannot handle PQ, HLG and SMPTE 428 so we are loading these as linear
images. Options have been added to choose the curve to export to for
floating point.

Chrome thinks PQ and SMPTE 428 are ok, but HLG has issues, and I don't
have the energy right now to debug that. I suspect I need to take into
account the nominal luminance value.