kis_paintop_box, improve icons and buttons size
replace the variable iconsize with a buttonsize, as it was actually used to set buttons sizes

add a new iconsize variable really used to set iconSize on buttons

also adjust accordingly the menuPadding variable used on mirror buttons to compensate
for their side arrow

TODO:
-it would be great to change ktoolbar to set all the other toolbar buttons size to 32
independently from selected theme, as it would harmonize perfectly the toolbar.

-maybe add a user setting to change the toolbar buttons size as discussed on
https://invent.kde.org/graphics/krita/-/merge_requests/679#note_188124
though it would need to change the size for toolbar buttons in both
ktoolbar and kis_paintop_box, and it would need to change both the
fixedSize and the iconSize (either with separate values, or doing
something like "iconsize = buttonsize - 10").