new smaller reload preset icon
add a 16px reload preset icon,
and use it in preset editor and popup palette.

Also use it in the Advanced color selector for "Create a list of colors from the image"

It's almost the same as updateColorize icon, except updateColorize has the special size
of 18px as needed for layer properties icons. So making a proper 16px version is better
for other places.