Add a new action to paste into the active layer
This adds a new entry named "Paste into Active Layer" to the Edit menu.
It pastes the image from the clipboard into the active layer or mask,
then activates the Transform tool, allowing the image to be positioned.

The layer is not erased under the pasted image, but the pasted image is
mixed in the same way as when a selection is moved.