new icons for the Artistic color selector, and a few other fixes
new icons for the Artistic color selector

change the Toggle gamut mask button to be disabled when no mask is set,
so it keeps showing the informative message "Select a mask in the Gamut Masks docker";
and only enable it to toggle to "Mask is disabled" when a mask is actually loaded.

new configure-thicker icon, used in several dockers.
I'm keeping the other thinner configure icon in a few places where it fits better.