fix all buttons and icons size issues in KisResourceItemChooser
fix all buttons and icons size issues in KisResourceItemChooser,
by porting viewModeButton from QToolButton to KisPopupButton

This improves a lot the visual aspect of the top bar in the preset selector popup and docker,
in the gradient popup and in the gamut mask docker.

Now all 3 buttons (Tags, Display and Bundles) are always the same height, and this
with any style selected (or at least both with Fusion and Breeze).

And it allows to get rid of the redundant arrow icon on the viewModeButton.