Fix MyPaint brushes sorting by engine in Brush Editor
Before this commit, MyPaint brushes would appear in all
categories in the item viewer docker just because
they have no 'paintop' metadata in the database.
This commit ensures that the proper metadata is created.

Note: you need to recreate a database to see the effect
if you already have MyPaint brushes there.