Fix switching presets on assigning tags
Before this commit, assigning a tag to the preset would switch
Brush Presets docker to that tag; and deassigning a tag would *not*
switch the tag, which resulted in a switch of presets.
This commit fixes this behaviour.

It contains the following changes:
- don't reset the model in KisTagFilterResourceProxyModel
unnecessarily
- switch to 'All' tag if you untag the resource from the current tag
- don't switch tags if you assign a tag
- remember the resource selected just before calling the context menu
and restore it afterwards (this fixes the resource not being selected
after deassigning a tag and switching to 'All').

Still existing issue: if you switch tags, the preset doesn't seem
to be selected in the item view even though it's still the same preset
active and even when it is available in the item view.