Fix values in columns in KisTagResourceModel
Before this commit, tags in context menu in the Brush Presets
docker would be in a different order than tags in the combobox.
Moreover even setting a proper sorting on the model wouldn't work.

This commit ensures there is proper sorting on all models the context
menu uses, and it ensures that the model returns correct data
when asked for a column (before, all QVariant for columns bigger than
StorageActive column from the ResourceModel would be returned Invalid
because they cannot be handled by the QueryMapper since they are
unique to TagResourceModel).