Add test for columns in TagResourceModel
Before this commit, TestTagResourceModel wouldn't test
roles besides the ones it defines itself (even though
here are plenty of roles that it takes from ResourceModel
via ResourceQueryMapper), and it wouldn't test columns
(what the model returns if instead of
data(index(i, 0), Qt::UserRole + role)
this is used:
data(index(i, role), Qt::UserRole)
which is important for sorting.

This commit ensures that the TestTagResourceModel tests both of those
things.