Add indexForResourceId() function to all resource models
Before this commit, there was no way to get an index and all other data
from the models knowing just the id (except for going through the whole
model and checking if the id fits), when you don't have the resource
loaded yet.
This commit adds indexForResourceId() function to make it easier for
other elements of the system to get the index and all other data
(since having an index means you have access to everything)
without having to write a loop just to find the correct index,
and without loading the resource.