Android: Fix floating dockers becoming unresponsive
In Qt's internal compositor API for Android, it added windows to
windowStack based on when they were created. For the most part this
worked. However in some cases Dock widget were created before the parent
widget. This resulted in all events going to MainWindow, even when dock
widget was used.

The cases where this happend: app was in pause and the SurfaceView was
totally torn down (rendered black). App is restarted with Dockers
floating.