Storyboard Docker: Improve stability of move scene undo/redo operations.
Due to the way keyframes were rearranged to match scene order,
there were no assurances that keyframe data would not change
under the image during operations via the GUI thread. This was
dangerous, and could cause rare instances of the GUI locking up
or (worst case scenario) a crash.

So some additional care was taken in the StoryboardDocker to
use KisProcessingApplicator for `moveRows` implementation.