Add nice showing of metadata (engine type)
Before this commit, Resource Manager would only have a space
for metadata but wouldn't show it.
This commit adds showing metadata.
Note that for paintoppresets it shows only the engine name,
for gamut masks only description,
and for everything else it shows pairs key-value,
but since nothing really uses the metadata (even gamut masks
descriptions are all empty in the default resources)
except for paintoppresets (for engine name), it's not a problem.