Always switch to the Select Shapes tool from the Edit Path tool
When the Select Shapes tool is active, it is currently possible to press
ENTER or double click to activate the Edit Path tool or the Text Tool.
The Text Tool always switches back to the Select Shapes tool when the
action is completed, however the Edit Path tool only does that if it was
activated from the Select Shapes tool. If it was activated from the tool
box, then pressing ENTER does nothing, and a double click on an inactive
area deselects all shapes.

This patch makes the behavior more consistent by changing the Edit Path
tool so that it switches to the Select Shapes tool in both conditions.