Remove the now unused temporary activation infrastructure within tools
This removes a lot of parameter passing that made tools aware of whether
they had been activated temporarily, as well as methods to request the
temporary activation of other tools.

If this awareness from within any tool has to be reintroduced at any
point, an alternative to parameter passing might be preferred. The tool
manager class is best suited to handle this state centrally, like it
already does for the temporary activation of the Line tool.