Fix autosave triggered only once per file
KisDocumentUndoStore's signal wasn't connected to the
KisImageSignalRouter, which caused the issue that
the autosave was only triggered once.
From the user point of view, it also resulted in the editing time
in File -> Document Information to be much lower than expected.
This commit connects the Document Undo Store to the SignalRouter
and ensures some more signals in KisDocumentUndoStore.
Those additional signals/emitions might not be needed.

BUG:434503
BUG:433513