Fix loading of embedded ASL patterns and gradients to the resource server
The ASL layer styles should not store the resources inside themselves.
Instead, they should upload the resources into the global resource server
and fetch them from there when needed. It fixes GUI elements to know
what resource is active in the layer, which actually has a cloned
layer style.

The patch introduces a special helper class KisLinkedResourceWrapper,
which remembers md5 and filename of the resource and then can be used
to fetch the resource from the server using this info.