3rdparty: disable autotools maintainer mode on gettext
In macOS, it seems that there's a filesystem race condition with
autotools's status check, which makes it trigger a rebuild of
gettext-runtime's config.h.in. I could only reproduce it with a manual
patch of the files, which does not update their timestamp.
Rebuilding config.h.in is not possible unless you have autotools
installed (and in CI they are certainly not), so I'll disable it
explicitly.

CCMAIL: kimageshop@kde.org