Fix crash on saving a huge image to .kra
Before this commit, an image of size bigger than 66k
in any dimension would crash Krita because it tries to
scale the projection to 256x256 image, which means that
the scale would be less than 1/256.
KisTransformWorker uses KisFixedPoint to scale,
which means that the scale would (in KisFixedPoint maths)
equal 0, which means there is division by 0.

BUG:432182