Ensure that transform worker won't try to scale to 0
Transform Worker did have a check for scale being 0...
but it's only in the double value. KisFilterWeightsBuffer
converts those values to KisFixedPoint, which has
much different precision (the scale cannot be lower than
1/256, otherwise it gets rounded to 0).
This commit adds a check after converting to KisFixedPoint.

Note: it would be best if the KisFixedPoint had higher precision.

I added the conversion check there, because it's an easy operation,
but it could be KisFilterWeightsBuffer and
KisFilterWeightsApplicator checking the values to make sure
they're correct transformation-wise and KisTransfromWorker only
checking some kind of buffer->valid() and applicator->valid() result.

CCBUG:432182