Make primary action and alternate action implementations the same
This allows the user to hold Shift to snap before or after pressing
their mouse button.

Programming-wise, this also means that the assistant tool doesn't take
advantage of the distinction between primary and alternate actions, so
it's up to the coder to manually check for modifiers() in the event object.