Fix saving metadata
KoXmlWriter is highly dangerous: if you pass a temporary
QByteArray created from a QString in the argument of a method,
the char* of the QByteArray will be deleted after the call to that
method has ended. Then the char* stored in KoXmlWriter to be
used as the end tag becomes invalid, and, in this case, likely to
be overwritten by the next QByteArray that get created from a
QString.