Fix KoQuaZipStore error checking
Before this commit, KoQuaZipStore would convert the number
of bytes written into a bool, and then if the bool was true
(which means if the number of bytes was higher than 0)
it would return the length of whole data as the number
of bytes written. That was the result of copying
the line across multiple commits, when the line was only
relevant a few commits ago in a KArchive (when writeData
was used and it did return bool for success).

Since now KoQuaZipStore uses write(), which returns the int,
this is the number that should be returned instead.

(See the deleted KArchive line 199 in commit 593dba2b70980).