Fix thumbnails for Resource Manager
Before this commit, thumbnails in the resource manager
on the right side of the dialog would render wrong
(for example when the palette thumbnail was very vertical
and it was drawn on the left side instead of in the middle).
This commit ensures that the same code
as for the resource item delegate thumbnails is used.