Change Import a bundle into Create bundle button
For now the button only opens a new create bundle dialog
without adding the already selected resources into the bundle,
but it can be added later.

Before this commit the button did nothing and was labeled
"Import a bundle", but because .asl and .abr files are
storages just like a bundle and it would be best if the user
didn't have to choose between two buttons but use just one
to import any kind of resource, resources or collection
of resources. Hence the Import a bundle button
became obsolete.