Bugfix: problems with transforming mirrored/flipped shapes
The handle (called `Direction` in constructor) we get from DefaultTool
is based on KoSelection of a shape. When we extract shape and operate on
it, the transformation applied on it map the handles incorrectly. This
doesn't happen until the shape is flipped where then bounding box of the
shape and KoSelection no longer map to same points (they map to the
opposites).