Bug 318347 grid brush keeps grid alignment when increasing brush size
This MR addresses a wishlist item to keep the grid alignment on the grid brush
when increasing the brush size. This introduces two new brush settings "Grid width"
and "Grid height" while removing the "scale" setting. This is to set exactly what
the size of each cell in the grid is instead of it being determined by the size of
the brush.

Test Plan
---------
1. Select a grid brush and draw a stroke
2. Increase brush size
3. Draw a new stroke. The ellipses in both strokes should keep alignment
   and be the same size

BUG:318347