Basic support for Video Animation Import
This patch allows importing video and animated images directly into
krita using ffprobe and ffmpeg. It is mostly based on scott's python
plugin, with a few extras, these include:

- Wider video and animated images support (gif works properly)
- Using the slower ffmpeg copy, it works even without ffprobe if needed
- loading the thumbnails via stdout pipe without the disk overhead
- ability to create new documents or import into current document
 * note: ffmpeg import will use default settings for its export of pngs
    which may not exactly honor the users choice of colorspace as this
    is "basic support".

This patch also includes an ffmpeg wrapper that is based on trying to
merge in the future, both the videosaver and the recorder wrapper into
one. The wrapper includes the ability to not only find ffmpeg/ffprobe,
but also to record the capabilities of the ffmpeg encoders and decoders.

This patch as a whole should allow for importing pretty much almost any
format ffmpeg supports. This of course means any format not bundled into
ffmpeg such as an animated webp decoder, or in the case of FOSS versions
of ffmpeg, h264 and h265 will not work. If the user does get a version
that does support it, the capabilities would be recorded and they would
be able to open those formats as well.