Crash when clicking Gradient Map/Palettize filter
This only happens with these two items because they both create a
QScroller which when used while Kinetic Scrolling being enabled is
pushed to a recognizers container. But when we delete the widget, the
cached value which was next to scroller's recognizer (a coincidence) is
taken to be valid hence a crash.

BUG:430802