Add KisStorageModel::importStorage
Before this commit, there was a function in KisResourceLocator
that could add a storage, but it had to be already in the resource
folder. Various dialogs and functions had to copy the storage file
into the folder themselves.

This commit adds a function that does both copying and adding
to the database.