Add ResourceImporter - initial implementation
Before this commit, DlgResourceManager didn't have any way to import
resources. This commit adds a class Resource Importer that is
used to import resources in DlgResourceManager.

Now the user can choose the resource type for every mimetype
of the files that can be used for multiple resource types
(like PNG or SVG).

Note: it's not possible yet to choose a different resource type
for different files with the same mimetype. However the user
can cancel the dialog, which would allow them to then
import for example only brush tips or only patterns separately.

Bundle import now will fail if you try to import a bundle
with the same name as a bundle already existing in the resource folder.