Fix range of Smudge Radius option
After the commit noted below the smudge radius value showed in
the GUI was multiplied by 100. That is, to get smudge radius
100%, the user had to select value 10000%.

Since the bug seems to have been introduced 6 years ago, I'm
not sure we can "just fix" it. So now the new presets will have
a 'SmudgeRadiusVersion' tag. If the tag is not found, then the
smudge radius value will be divided by 100.

The patch also introduces the logic of changing the range of
smudge radius. Old algorithm has range of 0...300%, and the
now one has a shorter range 0...100%. It happens because the
new algorithm uses the brush as weights for the samples, so we
cannot sample outside the brush bounds.

https://invent.kde.org/graphics/krita/-/commit/8d8e73359728ae778dd128da8d55993f640ca8d6